refactor: extract messages into a standalone openfactcheck.messages package
Move the chat message types into a top-level messages/ package that both
chat/ and prompts/ depend on, breaking the prompts -> chat edge. chat and
prompts are now standalone peers on the shared messages foundation. The
public import path is openfactcheck.messages; chat no longer re-exports it.
