+++ docs/reference/chat/index.md
@@ -8,7 +8,6 @@
 ## Pages
 
 - [`ChatClient` class](client.md) — the facade callers use to send messages and receive completions or streams.
-- [Messages](messages.md) — system, user, assistant, and tool messages that flow through the client.
 - [Requests](requests.md) — the bundled payload a backend receives (messages + config + runtime).
 - [Responses](responses.md) — complete responses, streaming events, and usage data.
 - [Config](config.md) — provider-specific configuration and runtime settings.
