+++ tests/chat/backends/openai/test_normalize.py
@@ -12,7 +12,7 @@ from openfactcheck.chat.errors import (
     ProviderError,
     RateLimitError,
 )
-from openfactcheck.chat.messages import AssistantMessage, SystemMessage, ToolMessage, UserMessage
+from openfactcheck.messages import AssistantMessage, SystemMessage, ToolMessage, UserMessage
 from openfactcheck.chat.responses import FinishReason
